remove cnn/fam model, to be replaced by single tfmodel
