--- grc/inspector_cnn.xml
@@ -1,48 +1,0 @@
-<?xml version="1.0"?>
-<block>
-	<name>CNN</name>
-	<key>inspector_cnn</key>
-	<category>[Inspector]/AMC</category>
-	<import>import inspector</import>
-	<make>inspector.cnn($typev.typ,$vlen,$graphfile)</make>
-
-	<param>
-		<name>Input Type</name>
-		<key>typev</key>
-		<type>enum</type>
-		<option>
-			<name>Message</name>
-			<key>message</key>
-			<opt>size:1</opt>
-			<opt>typ:"message"</opt>
-		</option>
-		<option>
-			<name>Complex</name>
-			<key>complex</key>
-			<opt>size:gr.sizeof_gr_complex</opt>
-			<opt>typ:"complex64"</opt>
-		</option>
-		</param>
-	<param>
-		<name>Vec Length</name>
-		<key>vlen</key>
-		<value>1</value>
-		<type>int</type>
-	</param>
-	<param>
-                <name>Graph file</name>
-                <key>graphfile</key>
-                <value></value>
-                <type>string</type>
-	</param>
-	<sink>
-    		<name>in</name>
-		<type>$typev</type>
-		<vlen>$vlen</vlen>  
-		<nports>$num_inputs</nports>
-	</sink>
- 	<source>
-    		<name>classification</name>
- 		<type>message</type>
-  	</source>
-</block>
