+++ python/CMakeLists.txt
@@ -32,8 +32,6 @@ GR_PYTHON_INSTALL(
     FILES
     __init__.py
     sync_prototype_cc.py
-    fam.py
-    cnn.py 
     tfmodel.py
     DESTINATION ${GR_PYTHON_DIR}/inspector
 )
